Added IDX_enter_if_input() to put input values on stack automatically
The former behavior required pressing "Enter" to move input values
from the input field onto the stack. This was inconsistent with the
typical RPN calculator operation where any operation key would put a
newly entered value on the stack automatically. This change makes the
manual operation more natural. This does not change running programs,
however, there may be a performance penalty for checking the HTML
values for every operation. This might be eliminated by noting key
presses or state changes and only performing the enter when it is
known that new values have been entered.